追加と更新ViewModelに通知メソッドがそれぞれあったのでManagerでもつ
