+++ QtlTodoList/Screen/TodoDetails/AddTodo/AddTodoCompletedButton/AddTodoCompletedButtonViewModel.swift
@@ -27,30 +27,4 @@ extension AddTodoCompletedButtonViewModel {
         title.wrappedValue = ""
         message.wrappedValue = ""
     }
-
-    func sendNotificationRequest(notificationDate: Date,
-                                 isNotification: Bool,
-                                 title: String,
-                                 message: String
-    ) async {
-        guard isNotification else { return }
-        let triggerDate = Calendar.current.dateComponents([.year, .month, .day, .hour, .minute], from: notificationDate)
-        print(triggerDate)
-        let trigger = UNCalendarNotificationTrigger(dateMatching: triggerDate, repeats: false)
-        let content = UNMutableNotificationContent()
-        content.title = title
-        content.body = message
-        content.sound = UNNotificationSound.defaultRingtone
-
-        let request = UNNotificationRequest.init(
-            identifier: "CalendorNotification",
-            content: content,
-            trigger: trigger)
-        let center = UNUserNotificationCenter.current()
-        do {
-            try await center.add(request)
-        } catch {
-            print(error.localizedDescription)
-        }
-    }
 }
